+++ app/src/test/java/org/hyperledger/besu/cli/BesuCommandTest.java
@@ -2553,6 +2553,28 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
+  public void txSenderNonceIndexEnabledOptionShouldWork() {
+    parseCommand("--tx-sender-nonce-index-enabled=true");
+    verify(mockControllerBuilder).senderNonceIndexingEnabled(booleanArgumentCaptor.capture());
+    verify(mockControllerBuilder).build();
+
+    assertThat(booleanArgumentCaptor.getValue()).isTrue();
+    assertThat(commandOutput.toString(UTF_8)).isEmpty();
+    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
+  }
+
+  @Test
+  public void txSenderNonceIndexDisabledOptionShouldWork() {
+    parseCommand("--tx-sender-nonce-index-enabled=false");
+    verify(mockControllerBuilder).senderNonceIndexingEnabled(booleanArgumentCaptor.capture());
+    verify(mockControllerBuilder).build();
+
+    assertThat(booleanArgumentCaptor.getValue()).isFalse();
+    assertThat(commandOutput.toString(UTF_8)).isEmpty();
+    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
+  }
+
+  @Test
   public void genesisStateHashCacheEnabledShouldWork() throws IOException {
     final Path genesisFile = createFakeGenesisFile(GENESIS_VALID_JSON);
     final ArgumentCaptor<EthNetworkConfig> networkArg =