+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/chain/Blockchain.java
@@ -14,6 +14,7 @@
  */
 package org.hyperledger.besu.ethereum.chain;
 
+import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.Hash;
 import org.hyperledger.besu.ethereum.core.Block;
 import org.hyperledger.besu.ethereum.core.BlockBody;
@@ -238,6 +239,15 @@ public interface Blockchain {
    * @return The transaction location associated with the corresponding hash.
    */
   Optional<TransactionLocation> getTransactionLocation(Hash transactionHash);
+
+  /**
+   * Returns the transaction hash for a given sender address and nonce.
+   *
+   * @param sender The sender address.
+   * @param nonce The transaction nonce.
+   * @return The transaction hash, if indexed.
+   */
+  Optional<Hash> getTransactionHashBySenderAndNonce(Address sender, long nonce);
 
   /**
    * Adds an observer that will get called when a new block is added.