+++ ethereum/referencetests/src/main/java/org/hyperledger/besu/ethereum/referencetests/ReferenceTestBlockchain.java
@@ -16,6 +16,7 @@ package org.hyperledger.besu.ethereum.referencetests;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
+import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.Hash;
 import org.hyperledger.besu.ethereum.chain.BlockAddedObserver;
 import org.hyperledger.besu.ethereum.chain.Blockchain;
@@ -125,6 +126,11 @@ public class ReferenceTestBlockchain implements Blockchain {
 
   @Override
   public Optional<TransactionLocation> getTransactionLocation(final Hash transactionHash) {
+    throw new NonDeterministicOperationException("Transaction location may be different on forks");
+  }
+
+  @Override
+  public Optional<Hash> getTransactionHashBySenderAndNonce(final Address sender, final long nonce) {
     throw new NonDeterministicOperationException("Transaction location may be different on forks");
   }